--- constants.py
@@ -1,5 +1,0 @@
-MATERIAL_REGISTRY_TABLE_PREFIX = 'MATERIAL_REGISTRY'
-MATERIAL_TABLE_PREFIX = "material_"
-MODEL_FILE_NAME = 'classifier.joblib'
-STAGE_NAME = "@ml_models"
-METRICS_TABLE = "CLASSIFIER_METRICS"
