+++ python_packages/profiles_classifier_pkg/MANIFEST.in
@@ -1,0 +1,3 @@
+include README.rst COPYING version.txt requirements.txt
+recursive-include docs *.rst *.png *.ico *.txt *.yaml
+graft profiles_classifier/config
