+++ python_packages/profiles_classifier_pkg/profiles_classifier/__init__.py
@@ -1,0 +1,5 @@
+from profiles_rudderstack.project import WhtProject
+from .model import ChurnModel
+
+def register_extensions(project: WhtProject):
+    project.register_model_type(ChurnModel)
