+++ python_packages/profiles_classifier_pkg/setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(
+    name='profiles_classifier',
+    version='1.0.0',
+    author='Ambuj Mishra',
+    author_email='ambujm@ruddersatck.com',
+    description='A py_native model package that implements the classification/regression model',
+    packages=['profiles_classifier'],
+    include_package_data=True,
+)
