+++ sample_profiles_project/models/profiles.yaml
@@ -1,0 +1,23 @@
+models:
+  - name: &model_name churn_predict_model
+    model_type: churn_prediction
+    model_spec:
+      inputs:
+        - packages/feature_table/models/shopify_user_features
+
+      train:
+        file_extension: .json
+        file_validity: 2m
+        config:
+          data: &model_data_configs
+            package_name: feature_table # Name of the package where the profiles feature table is defined (declared in pb_project.yaml packages)
+            label_column: is_churned_7_days
+            label_value: 1
+            prediction_horizon_days: 7
+            features_profiles_model: 'shopify_user_features'
+            # output_profiles_ml_model: *model_name
+            output_profiles_ml_model: 'shopify_churn_test'
+
+      predict:
+        config:
+          data: *model_data_configs
